--- src/main/java/edu/project2/MazeGenerator.java
@@ -1,5 +1,0 @@
-package edu.project2;
-
-public interface MazeGenerator {
-    Maze generateMaze();
-}
