--- src/main/java/edu/project2/PathFinder.java
@@ -1,7 +1,0 @@
-package edu.project2;
-
-import java.util.List;
-
-public interface PathFinder {
-    List<Cell> getPath();
-}
